Added implementation of task search by using qdrant vector database

+++ src/AI.Devs.Reloaded.API/Contracts/OpenAi/Embedding/Data.cs
@@ -1,3 +1,3 @@
 ﻿namespace AI.Devs.Reloaded.API.Contracts.OpenAi.Embedding;
 
-public sealed record Data(string @object, int index, List<double> embedding) { }
+public sealed record Data(string @object, int index, List<float> embedding) { }
